binary-file: fix asynchronous target writes
chunk_available() can start asynchronous write operations. We add a busy
flag to the target, which is !0 when the target itself is writing
data to flash. Before trying to write to the target we make sure it is not
busy. If it is, we can still decode some data and wait for the target to
become ready. We assume that bf_decoded_buf is free (i.e. data has been sent)
when chunk_available() returns, so we can reuse the same buffer to decode
some more data.

fix 689ad4395a10b8e028fe4e0387f5d0a9a0d4170b