http_urls: set fixed 32bytes size for urls data structures
url data structures are 16 bytes aligned and their size is 24bytes, this
made the search loop in find_url() fail (ptr was incremented by 24bytes
each iteration). We force size to be equal to 32bytes, so that alignment
is respected and find_url() get fixed.

There must be a proper way to do this, but I can't find it at the moment