fix: rx-method-http-lwip: introduce temporary errors for http requests
This was necessary for put() when there's no space enough for the
request's buffer to be appended to the binary file. In such case we
say that a temporary error has taken place and retry appending the buffer
when new space has been freed in the binary file's circular buffer.