binary-file: introduce handling of decoded buffer as a circular buffer
stk500 (optiboot bootloader) asks for fixed size chunks (i.e. chunks whose
size is equal to the size of a flash page, since non page aligned writes
are forbidden). The problem is that binary_format->decode_chunk can produce
a number of bytes which is not a multiple of chunk size.
As a consequence, we may not be able to write to flash the whole decoded
buffer, and a circular buffer becomes mandatory.
Note that actual decoded buffer size shall be an integer multiple of
chunk size, so that wrap arounds do not take place in the middle of a
flash page, and chunk_available can be kept simple.
Also note that flash page size shall be a power of two, but that should
always be the case.