MAIN HEADER: start adding filesystem support
Nordic files are transferred as ZIP's. Each ZIP file contains a .dat
file (which has to be sent as "init_packet" to the target) and one
or more actual firmware files. The problem is that we need to send
the .dat first, and then the firmware file(s), but the file order
inside the ZIP is arbitrary, so we have to temporarily store files
to persistent memory. We therefore introduce a "file container"
abstraction, which is something like a very simple filesystem without
directories.