jsmn: implement jsmn_search_token() and jsmn_scan_tokens()
The first function looks for a token with given name in a buffer.

The second function receives a pointer to an array of struct json_node
pointers, a pointer to an array of json tokens (previously setup by
jsmn_parse()), and a pointer to private data to be passed to nodes'
callbacks. The array of tokens is scanned according to "instructions"
contained in struct json_node's. Each time a token is examined, its
type is compared to the type field of the corresponding struct json_node.
If types don't match, the function returns error. If type is ok and token
is a string, its value is also compared with the one contained in the
corresponding struct json_node. If values don't match, the function returns
error. If everything is ok, the callback indicated by the current
struct json_node is invoked and child nodes are considered if present
(in such case recursion takes place).
If node has no children and the corresponding struct json_node requires
them, an error is returned. Otherwise the function returns the number of
successfully scanned tokens.

a pointer to s